status commit: backend hooked up to UI, but still shaking out errors
Right now a TCP error is firing (not really expected), and is cascading
into an unexpected situation where the IMAP protocol then tries to fire
its own error handler, but the handler we explicitly registered is not
there for some reason and then we get more errors.

It's very likely the TCP error is that my test-server is using a self-signed
cert.  Need to figure out how to deal with that...