Remove the move functionality in favor of higher level logic
The move functionality in imap.js tries to avoid expunging messages
that are not part of the move by removing their deleted flags
temporarily.  This has scaling implications and runs the risk of
clearing deleted flags.  More significantly, we don't support that
deletion model at all, so let's remove it.