status commit: IMAP appending tests, more loggest-izing
Deleting the pre-existing folder is now working, it's just hung on the
notification which is likely something dumb.

The obvious glitch right now is that we should probably tear down the
connection at the end of the test case because of the life-cycle issues
with the loggers.  I'm a little concerned about using real servers and
having them get angry about the rate at which we establish connections
given the potential high turnover of the unit test.  But that could just
be addressed by throttling everything (and allowing longer test runs)
when talking to real world servers that have a concept of abuse.