progress commit: changed sync time window calcs, tests fixed
This is an in-progress change to improve the time windows we use as we
sync further into the past.  The implementation thus far was somewhat
naive in that it would keep looking two weeks into the past if it
thought it could find more messages there.  It would do this when there
were deleted messages in the folder, because then we would expect more
messages to be there than our SEARCH with !DELETED would tell us about.
On localhost, the time window amusingly hit 1923 before I control-C'ed it!

There is an in-progress change to use a somewhat more adaptive strategy
where we:

- increase our time window's size if we don't find anything
- decrease our time window's size (and re-search) if we found way too many
  messages in the last search and the time range is large enough we can
  actually shrink the window (aka more than 1 day).
- if we have to decrease our window size, avoid increasing our window size
  until we clear the time region that caused us to reduce our size.  This
  will avoid oscillation.
- cap the size of the time windows based on how far we are in the past so
  as to avoid wastefully increasing the window size too much.  For example
  we want to avoid pathological window sizes if we're in an archive folder
  organized by year or the like.