adaptive / non-broken sync logic.
Finished implementing what was discussed in the last commit.  This does
not have unit tests, but it does not regress our existing unit tests.
I'll leave unit tests for tests of the logic and the imagined situations
that led to this seeming like a reasonable course of action.

NB: Yes, this can result in overkill scanning of a basically empty
folder that has non-expunged messages.