totally nuke db on upgrade, make fake acct not err on ops
Bumped the DB rev because previous versions were writing stuff to the
DB, but it's not what our current logic wants.  This was breaking the
email.gaia.visophyte.org example site if previously visited.

Updated fakeacct to know how to pretend it performed operations.  Flags
will not actually update; the UI is static.