fix db inconsistencies caused by failure to mark blocks dirty
The test_imap_internals.js 'account persistence' test was failing
because we would delete a header but not save the changes to the
block containing the header.  So when we restarted the account from
disk and synced the next time, we would say "oh, I see, this header
needs to be deleted", but we might also have already deleted the
body and managed to persist that or at least managed to alter the
block directory (which does get persisted) so as to break finding
the right block, etc.

This bumps the database revision, although I'm going to look at
quickly implementing the less-annoying upgrade where we re-create
the accounts.