debug support: deal with devicestorage strictness
DeviceStorage now only lets you write things that it thinks make sense
for its supported device storage types.  These are pictures, video, and
audio.  Unfortunately, JSON is not any of those things, so we are now
bad people and we pretend to be ".rm" files by appending ".json.rm"
instead of ".json".