fix TextEncoder/TextDecoder-in-core fallout, remove moot deps
Because TextEncoder/TextDecoder support landed, we don't require the
polyfill anymore.  TextEncoder/TextDecoder is available in an xpcshell
context too, so we are good.

The 'encoding' shim used by the e-mail library had some helpers to fix
non-standard-but-used encoding types, but it needed corrections in order
to work with the actual encodings supported by TextDecoder.

Unit tests are broken on the trunk, and this is a super-important thing
to land, so I am about to cherry-pick some fixes to make unit tests work
so I can land this.