support local_do/local_undo being async, re-green non-move tests
Prior semantics for local operations were that they were expected to be
synchronous or serialized one-shot manipulations with no need to hold a
mutex.  With moves, this is no longer a safe assumption.

Because the operation 'status' fields had the synchronous local stuff
baked in, this required breaking the status field into separate local
and server status fields.  I think this is generally cleaner.

I also reduced some of the complexity relating to operation state
tracking by eliminating the secret, untested "undoing an undone thing
is a redo".  We can support re-do, but we will want that to be explicit
and have tests.  This also included changing 'desire' to lifecycle with
a strict one-way series of transitions that terminate in 'undone'.

All IMAP unit tests are green except for the still in-progress move /
delete stuff.

I am going to update the ActiveSync operations now to fix any bit-rot.
I believe they are definitely missing some operations that should at
least be stubbed.