cache flushing, torture test, body size estimate fixes
This implements cache flushing with a green test.

It also introduces a torture test that expects the block purging and
thereby tests its hookup.

In the process of inspecting the torture test, I noticed that I screwed
up body sizing (the size var was being snapshotted before it reached its
final value) and that ActiveSync did not even try to compute the size!
So I moved the size computation into mailslice.

About to merge in trunk and then do a full test run, with any fallout
being fixed.