Add an explicit build target and have the test targets depend on it.
When trying to assist vingtetun in running the tests today, I realized
that the unit tests don't depend on our build process, which is bad.
Also, during the introduction of install-into-gaia, we lost our
gaia-email-opt.js build target which was our only "build but don't
install" target, and it's very handy to have one of those.