Bug 822882 - Initial bisect sync fix try #2.
makeDaysBefore's use of makeDaysAgo was foolish since it caused unexpected
rounding up.  Because of the various deadlines and pressures, I'm sort of
running around with my head cut off, so this wasn't a tremendously surprising
oversight.

This includes a timezone stopgap fix; I'm probably going to have to implement
the IMAP fake server soon for other horrible things that are going on, which
should eliminate the need for this hack.