Allow running with mozilla-central b2g desktop, and remove manual GC as it can cause the tests to hang.
The hang only happens in certain configurations, and seems to be racy. A 2012 MacBook Air reliably gets the hang though. With this change, with b2g18, there can be a segmentation fault that can occur. If the tests are broken up into separate runs then it should work though. mozilla-central seems to reliably complete all the tests though.