Bug 1024660 - persistent failure in test_mutation for pop3:fake
At some point something changed in b2g's execution (ex: this did not used
to fail, but when I upgraded my b2g it now consistently fails) so that ugly
leftover state in _TEST_pendingHeaderDeletes from a test step that was not
relevant to POP3 was failing due to a refusal to structured clone the state.

I'm not sure exactly what the change was and I assert we really don't need
to care since we don't want to be persisting the state in question anyways.

Note that bug 921050 will moot the structured clone problem as described in
the comment in the patch.