Bug 1039007 - [email/backend] Do not invoke operation callbacks until after any associated saves have safely hit the disk
The core of this is the save change.  This also exacerbated an intermittent
failure in test_activesync_recreate so I cleaned that up a little which
involved adding some more logging in this area.  (Well, I probably could
have gotten by without the extra logging, but I believe we are imminently
going to improve our upgrade mechanism, and that will want the logging, so
here we are.)