Bug 800402 - Update tests to catch the front-end regression, improve notification consistency because of move/deletion bug caught
expect_unread now also checks that the front-end state is properly updated.
For some reason I thought we had that base covered already.

As a side-effect of fixing that the mutation tests related to moves/deletion
caught that we weren't notifying in all cases.  So I simplified things so that
we always notify when the folder saves (because it's dirty).

This then caught that the upgradeDB job wasn't issuing a save or dirtying, so
I fixed that.  Hooray tests!