Bug 951312 - [email/backend] check_syncFolderList's doneCallback('idempotent') should be doneCallback(null, 'coherent-notyet') so we retry synchronizing the folder list
In bug 951076 we saw a syncFolderList failure from ActiveSync.
ActiveSync triggers an 'aborted-retry' which has us run a check
operation.  The check operation intends to have us retry by returning
'idempotent', but screws up and returns it as an error code instead of
the return value.

Also, idempotent is semantically the wrong thing for us to be returning.
'coherent-notyet' is better.