Bug 1091295 - [Email] Failure to normalize newlines in composed messages. May cause some SMTP servers (ex: qmail) to refuse to send mail with a 451 error.
Note that r+-ing this is an implicit r+ of
https://github.com/mozilla-b2g/mail-fakeservers/pull/24
which I have speculatively already landed.  In that vein, note that I've also
updated our package.json to NOT use a tilde for mail-fakeservers since the
tilde regrettably defeats my whole concept of "we don't take the upgrade
without a gelam upgrade".