Bug 1096025 - Make sent messages marked as read.
For IMAP we had a regression when we moved to email.js.  The imap.js impl we
were previously using would prepend the backslash for us.  browserbox does not,
which makes more sense, but we forgot to update our code.  This is now
corrected and checks have been added to test_compose to verify this.  (The
tests fail without the change.)

It also turned out that our POP3 implementation was not setting the messages as
read, so I've done that.

We also had the same problem for ActiveSync, but that was due to the server not
marking the message as read when saving to the sent folder.  (AKA it wouldn't
have been a problem in the real world).  I've updated this in mail-fakeservers
in the 0.0.33 release.