Bug 1128285 - Avoid snippet fetch hang by not leaking latches
Multipart bodies with N parts that we are interested in where the first N-1
body parts are sufficient to push us over our byte limit will result in us
creating and leaking allback latches for the succeeding parts.  This is
corrected by not creating the latch until we're sure we're going to try and
process the part.

This includes a fix and added test coverage in test_imap_parallelfetch that
fails without the fix.  I chose the specific test because it is specifically
a test of the snippet logic and intentionally tries to do things in parallel.
As such, it's the best place to also try and force a hang due to weird
concurrency.

(The other place I was considering to put the test was test_mime which also
has various MIME hierarchy permutations and snippet-checking logic.  However,
test_mime is more concerned about the correctness of consuming the various
structures rather than the mechanics of how we go about fetching things from
the server.)