Bug 825318 - Support registering downloads with the download manager. r=jrburke
Back-end support for registering with the mozDownloadManager web API and its
adoptDownload method.

This also includes some minor testing changes to allow the SOLO_FILE argument
to "make post-one-test" to keep and filter using the file extension so that
when doing SOLO_FILE=test_compose.js we don't also run files like
test_compose_blob.js and test_compose_detach.js.  This allows for faster
testing and less confusing reporting.  (We would run the other tests but not
report them because of the way the "post" part of the step ran.)

This also includes a fix for Bug 1111724 - Handle text/* attachments correctly.