Bug 1083355 - Fix message sequence number off-by-one ordinal/count error
We needed to add one to the EXISTS count to be referring to the next allocated
message sequence number.  (UIDNEXT already had these semantics built-in.)

Prior to this fix we would break if there were no messages in the target folder
when falling back to using message sequence numbers.  And in all cases we would
slightly inefficiently get the message-id for at least one extra message that
couldn't possibly be the one we cared about.

I'm also removing a line that was obsoleted in bug 1116694 (which added this
message sequence number path), but was not removed.

I've updated mail-fakeservers to explode if it sees a 0 (or negative) message
sequence number or UID.