non-backwards-compatible: use '.' instead of '/' for id delimiters
Now that we're potentially surfacing our various id's in URLs that may look
path-like, '/' is arguably a bad choice for our id delimiter.  I've chosen '.'
to replace it because it doesn't get URI encoded and has the same sorting
characteristics as '/' since it precedes it (and ASCII numerals).

There were other options, but none were clearly superior so I just ran with
this.  We can change things again in the future if desired.