The variant is stored on the specific tests.
The logic was accepting the first variant it found for any test type because
it was performing comparisons against properties that didn't exist and
undefined === undefined.  This fix looks at the first test for the summary.
I think this is safe in general because separate variants get separate test
runs and separate JSON files and so all the tests in a file will have the
same variant.  (There just may be multiple tests run in that single file.)

Probably the more correct solution is to include the variant at the location.
But I'm not going to do that tonight, so I'll publish this PR for now,
hopefully not landing this, but maybe landing it and fixing it later.