Bug 1160304 - Fix test_storage_upgrade.js ordering
I had already fixed a similar problem for other uses of expect_unread, but this
seemed to have recently been pushed into being a reliable failure.

There was also a missing "var" declaration in a for loop in the logic the test
tests that was not directly involved in this failure, but which is a good idea
to fix.