Stream attachment downloads.
IMAP downloads attachments one chunk (currently 1MB) at a time using
FETCH ranges. POP3 just blasts it through to disk (in chunks). On ActiveSync,
moz-chunked-arraybuffer lets us handle data as we receive it. (This also
entailed moving to their "multipart" downloading response format, not to
be confused with MIME multipart.)

Attachments now flow to IndexedDB in chunks, stored as `{ parts: [...] }`,
where each part is a partial Blob of the incoming file. When the file has
finished downloading, we concatenate all the parts into one super-Blob.

If download fails, we implement `check_download`, telling it to 'moot' the
operation, so that we don't attempt to repeatedly download a troublesome file.

This uses the WHATWG Streams reference implementation with one minor tweak,
as noted in the updated README. While their reference implementation is still
in flux as the standard is being developed, our use is isolated to a few files,
which should be no problem to update as the spec becomes finalized.

The Streams implementation supports dealing with backpressure, however this
patch does not address backpressure. IMAP is the only protocol we could do
anything about without socket-level support.

Due to performance and implementation concerns about whiteout-io's MimeParser,
this patch converts most of our MIME parsing to use jcranmer's jsmime, which was
fortunately relatively simple, modulo a bunch of refactoring in `imapchew.js`.

We still, at the moment, use whiteout-io's MimeParser for parsing text bodies.
Swapping that over could potentially be a followup bug.

I noticed that ActiveSync's attachment handling is somewhat broken (both with
and without this patch):

https://bugzilla.mozilla.org/show_bug.cgi?id=1182955