checkpoint prior to renaming activesync/protocol
Unfortunately, the explicit manual mapping for activesync/protocol for jsas
means that us having an activesync/protocol directory that's effectively
mapped at the top level turns out quite poorly.  So activesync now gets a
"smotocol" directory.  This is not permanent because although it's somewhat
amusing, it's also extremely and unacceptably idiomatic.  I'm doing it at
all because I'm admittedly a little punchy right now and I want to be
compelled to right this error at some point.