fix deeply concerning quoted reply regression
We were missing newlines between quoting boundaries.  Maybe a bug 1169130
regression?  (Noting that I hadn't landed that fix on the trunk, which is
good.)  As I posit on that bug, it could be a combination of that fix
having eaten the "wrote:" trailing newline and then my single-line reply
not including a trailing newline.

Since I think we canonicalize to remove trailing newlines, I believe this is
a proper fix, but there is almost certainly a need for improved test coverage
and investigation here.