normalize _sendPromisedRequest, cleanup account creation
_sendPromisedRequest-replies now always include an explicit "data" property
which is what the promise gets resolved as.  (The implementation went through
a few spurts, but not all constraints were propagated...)

Account creation was suffering from its assumptions around this API.
learnAboutAccount and tryToCreateAccount should now consistently return values,
with tryToCreateAccount working regardless of whether domainInfo/configInfo is
passed or not.

The account manager will not automatically instantiate the account in order to
trigger syncFolderList.  This assumes a connection is passed-through; a cleanup
in the future would be advisable.