logic-inspector timeline view progress checkpoint
Need to get back to development proper.  This includes a lot of speculative
documentation and configuration files for my plan for a timeline based view.
It's not implemented yet.

The current state, however, does work and enables the "href" to be pointed
at a file ending in ".log" which is assumed to be a text-based log with
lines that look like "logic: {JSON}\n" and it can extract and display them
in a continuous vertical list.  The next convoy commit will include these
changes.