WIP support for messages-by-folder
Not yet consumed yet, so there may be a gap or two.

So far the possible views have been conversation-centric, wherein
we can view conversations-by-folder and messages-by-conversation, but
there was no messages-by-folder.

In theory the query subsystem could provide a means for this, but
without a materialized index over the message timestamps, this would
not be remotely efficient.

This change introduced the `msgIdsByfolder` which is a twin of
`convIdsByFolder` but without the `QuantizedHeight` payload which exists
for clever virtual list purposes but is not currently used by
glodastrophe.