Move all account types to js/accounts/, more import fixes to come.
The original account types living in the root was never particularly
intentional and does not scale.  The "extras" dir made sense in a
multi-consumer implementation where it was important to delineate
the core, supported accounts from random experimental ones.

So I'm moving all the accounts to js/accounts/, which does not imply
any level of support, but keeps like things together.

VS Code made some automated fix-ups and I fixed engine_glue, but there
are a bunch of imports that will need an extra layer of "../" prepended
for the original core accounts that lived directly under js/.  Those
are coming next.